refactor: make custom GrpcSslContext into GrpcSslContextFactory
