--- spring-cloud-gateway-server/src/main/java/org/springframework/cloud/gateway/config/GRPCSSLContext.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2013-2020 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.cloud.gateway.config;
-
-/**
- * @author Alberto C. Ríos
- */
-public class GRPCSSLContext {
-// TODO remove this class, or make it a true factory that returns SSlContext
-	public HttpClientProperties.Ssl getSslProperties() {
-		return sslProperties;
-	}
-
-	private final HttpClientProperties.Ssl sslProperties;
-
-	public GRPCSSLContext(HttpClientProperties properties) {
-		sslProperties = properties.getSsl();
-	}
-}
